Handle comment reviews as a review
Many people just leave a comment instead of requesting changes. Today
those are ignored, but it leaves the PR labels in an odd state.

This starts treating comment reviews as being reviewed. It doesn't
request changes on the author because it's unclear who needs to take
action to move things forward.

Link: https://community.theforeman.org/t/reviewing-prprocessors-review-labels/38844